output: merge common lighting functions
This moves common lighting functions to TRX. This temporarily
introduces some helper functions to access TR2 specifics, like shade
tables, while the module remains split. These can later be eliminated.
