+++ src/libtrx/include/libtrx/game/output/const.h
@@ -7,3 +7,5 @@
     #define MAX_OBJECT_TEXTURES 2048
     #define MAX_SPRITE_TEXTURES 512
 #endif
+
+#define WIBBLE_SIZE 32
