+++ src/tr1/game/output.h
@@ -33,10 +33,6 @@ void Output_EndScene(void);
 void Output_DrawBlack(void);
 void Output_ClearDepthBuffer(void);
 
-void Output_CalculateLight(XYZ_32 pos, int16_t room_num);
-void Output_CalculateStaticLight(int16_t adder);
-void Output_CalculateObjectLighting(const ITEM *item, const BOUNDS_16 *bounds);
-
 void Output_DrawObjectMesh(const OBJECT_MESH *mesh, int32_t clip);
 void Output_DrawObjectMesh_I(const OBJECT_MESH *mesh, int32_t clip);
