+++ src/tr2/game/output.h
@@ -80,14 +80,6 @@ void Output_InsertShadow(
 
 void Output_CalculateWibbleTable(void);
 int32_t Output_GetObjectBounds(const BOUNDS_16 *bounds);
-void Output_CalculateLight(XYZ_32 pos, int16_t room_num);
-void Output_CalculateStaticLight(int16_t adder);
-void Output_CalculateStaticMeshLight(XYZ_32 pos, SHADE shade, const ROOM *room);
-void Output_CalculateObjectLighting(const ITEM *item, const BOUNDS_16 *bounds);
-void Output_LightRoom(ROOM *room);
 void Output_SetupBelowWater(bool is_underwater);
 void Output_SetupAboveWater(bool is_underwater);
 void Output_AnimateTextures(int32_t ticks);
-
-void Output_ResetDynamicLights(void);
-void Output_AddDynamicLight(XYZ_32 pos, int32_t intensity, int32_t falloff);
